+++ src/ui/react/settings/app.tsx
@@ -8,6 +8,7 @@ import { newSelectors } from "./modules/shared/store/app/reducer.ts";
 
 import { Routing } from "./router.tsx";
 import { ThumbnailGeneratorModal } from "./components/ThumbnailGeneratorModal/index.tsx";
+import { WelcomeModal } from "./components/WelcomeModal/infra.tsx";
 
 export function App() {
   const isDarkMode = useDarkMode();
@@ -35,6 +36,7 @@ export function App() {
     >
       <Routing />
       <ThumbnailGeneratorModal />
+      <WelcomeModal />
     </ConfigProvider>
   );
 }